% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalplot.R
\name{normalplot}
\alias{normalplot}
\title{Plot a normal distribution}
\usage{
normalplot(mu = 1, sd = 1)
}
\arguments{
\item{mu}{is the mean of the distribution (default = 1)}

\item{sd}{is the standard deviation of the distribution (default = 1)}
}
\value{
returns a plot of the normal distribution
}
\description{
Plot a normal distribution
}
\examples{
normalplot(10, 3)
normalplot(2, 0.5)
}
